backend/pyproject.toml
[project]
name = "socratiq-backend"
version = "0.1.0"
description = "Socratiq - AI-driven adaptive learning system backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "alembic",
    "pgvector",

    # LLM SDKs
    "anthropic>=0.40",
    "codex-app-server-sdk==0.3.2",
    "openai>=1.50",

    # Async tasks
    "celery[redis]",
    "redis",

    # Security
    "cryptography",

    # Utils
    "pydantic>=2.0",
    "pydantic-settings",
    "httpx",

    # Content extraction (Sub-project B)
    "bilibili-api-python>=16.0",
    "pymupdf>=1.24",
    "python-multipart",
    "youtube-transcript-api>=0.6.1",
    "yt-dlp>=2024.1",

    # Agent + SSE (Sub-project C)
    "sse-starlette",
    "numpy",

    # Auth
    "PyJWT>=2.8",
    "bcrypt>=4.0",
    "google-auth>=2.0",

    # Observability
    "structlog>=24.0",
    "sentry-sdk[fastapi]>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "testcontainers[postgres]",
    "fakeredis[lua]",
]
whisper = [
    "openai-whisper>=20231117",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
